+++ serialisation/src/main/java/org/ehrbase/serialisation/flatencoding/std/marshal/StdFromCompositionWalker.java
@@ -96,9 +96,7 @@ public class StdFromCompositionWalker extends FromCompositionWalker<Map<String, 
 
   @Override
   protected void handleDVText(
-      WebTemplateNode currentNode,
-      Map<String, List<WebTemplateNode>> choices,
-      List<WebTemplateNode> children) {
+      WebTemplateNode currentNode) {
     if (currentNode.getRmType().equals("ELEMENT")) {
       List<WebTemplateNode> trueChildren =
           currentNode.getChildren().stream()
@@ -108,17 +106,17 @@ public class StdFromCompositionWalker extends FromCompositionWalker<Map<String, 
                           || !n.isNullable())
               .collect(Collectors.toList());
       if (trueChildren.stream()
-              .map(WebTemplateNode::getRmType)
+              .map(WebTemplateNode::getId)
               .collect(Collectors.toList())
-              .containsAll(List.of("DV_TEXT", DV_CODED_TEXT))
+              .containsAll(List.of("coded_text_value", "text_value"))
           && currentNode.getChoicesInChildren().size() > 0
           && trueChildren.size() == 2) {
-        handleDVTextInternal(currentNode, choices, children);
+        handleDVTextInternal(currentNode);
       } else {
-        super.handleDVText(currentNode, choices, children);
+        super.handleDVText(currentNode);
       }
     } else {
-      super.handleDVText(currentNode, choices, children);
+      super.handleDVText(currentNode);
     }
   }
 }